+++ src/common/types/index.ts
@@ -1,3 +1,2 @@
-export * from "./type.js";
-export * from "./typeWrappers.js";
+export * from "./types.js";
 export * from "./parser.js";